+++ src/components/BirdDescriptionBlock/BirdDescriptionBlock.module.css
@@ -1,23 +1,31 @@
 .container {
   width: 40%;
-  height: 300px;
+  height: 23rem;
   display: flex;
   flex-direction: column;
   align-items: center;
+  justify-content: center;
   background-color: #444;
-  padding: 15px;
+  gap: 0.5rem;
+  padding: 8px;
   border-radius: 15px;
 }
 
 .topContent {
   width: 100%;
   display: flex;
-  justify-content: space-between;
+  justify-content: space-around;
+  align-items: center;
+}
+
+.nameBlock {
+  display: flex;
+  flex-direction: column;
 }
 
 .img {
-  width: 220px;
-  height: 170px;
+  width: 13rem;
+  height: 10rem;
   border-radius: 5px;
   object-fit: contain;
 }
