+++ src/components/ListOfAnswers/ListOfAnswers.module.css
@@ -15,14 +15,3 @@
   padding: 0;
   margin: 0;
 }
-
-.actionBlock {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.count {
-  font-size: 2rem;
-  color: white;
-}
