+++ src/components/ListOfBirdSpecies/ListOfBirdSpecies.module.css
@@ -6,6 +6,9 @@
 
 .birdSpecies {
   width: 25%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   padding: 12px;
   font-size: 1.2rem;
   text-align: center;
@@ -19,6 +22,9 @@
 
 .birdSpeciesActive {
   width: 25%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   padding: 12px;
   font-size: 1.2rem;
   text-align: center;
