+++ src/index.css
@@ -42,6 +42,10 @@ a {
   color: #fff;
 }
 
+p {
+  margin: 0;
+}
+
 a {
   font-size: 1.2rem;
   text-decoration: none;
@@ -82,8 +86,11 @@ button:disabled {
   color: #008966;
 }
 
-@media (max-width: 769px) {
+@media (max-width: 1025px) {
   html {
-    font-size: 12px;
+    font-size: 14px;
   }
+}
+
+@media (max-width: 769px) {
 }
